refactor(core): flatten dialect model and address review feedback
Promote the nested dialect types out of the @Value.Enclosing Dialect
holder into top-level classes in io.substrait.dialect, and rename
DialectDocument to Dialect (now carrying the builder/load/toYaml factory
methods). This removes the io.substrait.dialect.Dialect.DialectDocument
repetition.

Also addresses the remaining review comments:
- share a single package-scoped ObjectMapper instead of creating one per
  call, and have the union (de)serializers use it directly rather than
  casting JsonParser#getCodec()
- read directly from InputStream/File instead of slurping into a String;
  the InputStream overload no longer closes a caller-owned stream
- accept a single scalar as a one-element list in readEnums/readStrings,
  consistent with ACCEPT_SINGLE_VALUE_AS_ARRAY
- reject metadata on EXECUTION_CONTEXT_VARIABLE and forbid setting both
  writeTypes and ddlWriteTypes (they share the write_types field), so
  entries always round-trip
- make the (de)serializer classes package-private
- use InputStream.readAllBytes() in the spec-fixtures test
