+++ core/src/main/java/io/substrait/dialect/CastFailureOption.java
@@ -1,0 +1,7 @@
+package io.substrait.dialect;
+
+/** Permissible failure options for {@code CAST} expressions. */
+public enum CastFailureOption {
+  RETURN_NULL,
+  THROW_EXCEPTION
+}
