+++ core/src/main/java/io/substrait/dialect/DdlWriteType.java
@@ -1,0 +1,7 @@
+package io.substrait.dialect;
+
+/** Operable object types for {@code DDL} relations. */
+public enum DdlWriteType {
+  NAMED_OBJECT,
+  EXTENSION_OBJECT
+}
