+++ core/src/main/java/io/substrait/dialect/ExchangeKind.java
@@ -1,0 +1,10 @@
+package io.substrait.dialect;
+
+/** Exchange kinds for {@code EXCHANGE} relations. */
+public enum ExchangeKind {
+  SCATTER_BY_FIELDS,
+  SINGLE_TARGET,
+  MULTI_TARGET,
+  ROUND_ROBIN,
+  BROADCAST
+}
