+++ core/src/main/java/io/substrait/dialect/ExpandFieldType.java
@@ -1,0 +1,7 @@
+package io.substrait.dialect;
+
+/** Field types for {@code EXPAND} relations. */
+public enum ExpandFieldType {
+  SWITCHING_FIELD,
+  CONSTANT_FIELD
+}
