+++ core/src/main/java/io/substrait/dialect/NestedType.java
@@ -1,0 +1,8 @@
+package io.substrait.dialect;
+
+/** Nested types for {@code NESTED} expressions. */
+public enum NestedType {
+  STRUCT,
+  LIST,
+  MAP
+}
