+++ core/src/main/java/io/substrait/dialect/Notation.java
@@ -1,0 +1,9 @@
+package io.substrait.dialect;
+
+/** How a system function is written in the system's own syntax. */
+public enum Notation {
+  INFIX,
+  POSTFIX,
+  PREFIX,
+  FUNCTION
+}
