+++ core/src/main/java/io/substrait/dialect/ReadType.java
@@ -1,0 +1,10 @@
+package io.substrait.dialect;
+
+/** Read types for {@code READ} relations. */
+public enum ReadType {
+  VIRTUAL_TABLE,
+  LOCAL_FILES,
+  NAMED_TABLE,
+  EXTENSION_TABLE,
+  ICEBERG_TABLE
+}
