+++ core/src/main/java/io/substrait/dialect/SubqueryType.java
@@ -1,0 +1,9 @@
+package io.substrait.dialect;
+
+/** Subquery types for {@code SUBQUERY} expressions. */
+public enum SubqueryType {
+  SCALAR,
+  IN_PREDICATE,
+  SET_PREDICATE,
+  SET_COMPARISON
+}
