+++ core/src/main/java/io/substrait/dialect/SupportedExpressionDeserializer.java
@@ -4,8 +4,6 @@ import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonNode;
-import io.substrait.dialect.Dialect.ExpressionKind;
-import io.substrait.dialect.Dialect.SupportedExpression;
 import java.io.IOException;
 import java.util.Map;
 
@@ -14,7 +12,7 @@ import java.util.Map;
  * {@code LITERAL}) or a configuration object (e.g. {@code {expression: CAST, failure_options:
  * [RETURN_NULL]}}).
  */
-public class SupportedExpressionDeserializer extends JsonDeserializer<SupportedExpression> {
+class SupportedExpressionDeserializer extends JsonDeserializer<SupportedExpression> {
 
   @Override
   public SupportedExpression deserialize(JsonParser p, DeserializationContext ctxt)
@@ -25,21 +23,20 @@ public class SupportedExpressionDeserializer extends JsonDeserializer<SupportedE
     }
 
     ExpressionKind kind = ExpressionKind.valueOf(node.get("expression").asText());
-    ImmutableDialect.SupportedExpression.Builder builder =
-        SupportedExpression.builder().expression(kind);
+    ImmutableSupportedExpression.Builder builder = SupportedExpression.builder().expression(kind);
 
-    Map<String, Object> metadata = DialectJsonSupport.readMetadata(p, node);
+    Map<String, Object> metadata = DialectJsonSupport.readMetadata(node);
     if (metadata != null) {
       builder.metadata(metadata);
     }
     builder.addAllFailureOptions(
-        DialectJsonSupport.readEnums(node.get("failure_options"), Dialect.CastFailureOption.class));
+        DialectJsonSupport.readEnums(node.get("failure_options"), CastFailureOption.class));
     builder.addAllSubqueryTypes(
-        DialectJsonSupport.readEnums(node.get("subquery_types"), Dialect.SubqueryType.class));
+        DialectJsonSupport.readEnums(node.get("subquery_types"), SubqueryType.class));
     builder.addAllNestedTypes(
-        DialectJsonSupport.readEnums(node.get("nested_types"), Dialect.NestedType.class));
+        DialectJsonSupport.readEnums(node.get("nested_types"), NestedType.class));
     builder.addAllVariableTypes(
-        DialectJsonSupport.readEnums(node.get("variable_types"), Dialect.VariableType.class));
+        DialectJsonSupport.readEnums(node.get("variable_types"), VariableType.class));
 
     return builder.build();
   }
