+++ core/src/main/java/io/substrait/dialect/VariableEvaluationMode.java
@@ -1,0 +1,7 @@
+package io.substrait.dialect;
+
+/** When execution context variables are evaluated. */
+public enum VariableEvaluationMode {
+  PER_PLAN,
+  PER_RECORD
+}
