+++ core/src/main/java/io/substrait/dialect/VariableType.java
@@ -1,0 +1,8 @@
+package io.substrait.dialect;
+
+/** Variable types for {@code EXECUTION_CONTEXT_VARIABLE} expressions. */
+public enum VariableType {
+  CURRENT_TIMESTAMP,
+  CURRENT_TIMEZONE,
+  CURRENT_DATE
+}
