+++ core/src/main/java/io/substrait/dialect/WriteType.java
@@ -1,0 +1,7 @@
+package io.substrait.dialect;
+
+/** Write types for {@code WRITE} relations. */
+public enum WriteType {
+  NAMED_TABLE,
+  EXTENSION_TABLE
+}
