+++ core/src/test/java/io/substrait/dialect/SpecDialectFixturesTest.java
@@ -4,13 +4,11 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.networknt.schema.Error;
-import io.substrait.dialect.Dialect.DialectDocument;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.UncheckedIOException;
 import java.nio.charset.StandardCharsets;
 import java.util.List;
-import java.util.Scanner;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
@@ -26,10 +24,7 @@ class SpecDialectFixturesTest {
       if (stream == null) {
         throw new IllegalStateException("Fixture not found on classpath: " + resourcePath);
       }
-      try (Scanner scanner = new Scanner(stream, StandardCharsets.UTF_8.name())) {
-        scanner.useDelimiter("\\A");
-        return scanner.hasNext() ? scanner.next() : "";
-      }
+      return new String(stream.readAllBytes(), StandardCharsets.UTF_8);
     } catch (IOException e) {
       throw new UncheckedIOException(e);
     }
@@ -53,7 +48,7 @@ class SpecDialectFixturesTest {
     assertTrue(fixtureErrors.isEmpty(), () -> fixture + " is not schema-valid: " + fixtureErrors);
 
     // Parse, re-serialize, and confirm the output is still schema-valid.
-    DialectDocument parsed = Dialect.loadResource(resourcePath);
+    Dialect parsed = Dialect.loadResource(resourcePath);
     String reserialized = Dialect.toYaml(parsed);
     List<Error> roundTripErrors = SchemaValidator.validate(reserialized);
     assertTrue(
